requirements.txt
python-telegram-bot==20.7
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth==2.26.0
requests==2.31.0
httpx==0.25.1
httpx[socks]==0.25.1